tss: point KERNEL_TSS_STACK at the real boot stack; stop truncating RSP0
KERNEL_TSS_STACK was 0xFFFF8000FFFFF000ULL — a high canonical address that
is never mapped in this 1MB-linked kernel, used as the fallback/initial TSS
RSP0/ESP0. It had two problems:

* As a 64-bit constant assigned to the legacy 32-bit TSS esp0 field
  (gdt.c) it overflowed uint32_t (-Woverflow), truncating to 0xFFFFF000.
* If any path ever fell back to it, RSP0 would point at unmapped memory and
  the next ring3->ring0 transition would triple-fault.

Redefine it as the real boot stack symbol ((uintptr_t)stack_top), which the
boot code installs as the initial RSP/ESP and is therefore always mapped and
low (0x14a000). This is valid on both 32- and 64-bit, and being non-constant
it no longer overflows the 32-bit field. Export stack_top from multiboot.S so
C can link it.

Also fix the same latent truncation at the four sites that loaded a 64-bit
kernel_stack_top into a uint32_t before set_tss_kernel_stack() (scheduler.c
x3, syscall.c x1): widen to uintptr_t. Lossless today (kernel stacks are
<4GB) but the truncation was wrong in principle.

Default + MINIMAL_SECURE builds clean (warning gone); 14 rust tests pass.

+++ src/kernel/scheduler.c
@@ -242,7 +242,7 @@ void context_switch(int next) {
 
     set_current_task(next);
 
-    uint32_t kstack_top = tasks[next].kernel_stack_top;
+    uintptr_t kstack_top = tasks[next].kernel_stack_top;
     if (kstack_top == 0) {
         kstack_top = (addr_t)&kernel_stacks[next][KERNEL_STACK_SIZE - 16];
     }
@@ -276,7 +276,7 @@ void schedule(void) {
     }
 
     if (next != cur && tasks[next].cr3 != 0 && next != 0) {
-        uint32_t kstack_top = tasks[next].kernel_stack_top;
+        uintptr_t kstack_top = tasks[next].kernel_stack_top;
         if (kstack_top == 0) {
             kstack_top = (addr_t)&kernel_stacks[next][KERNEL_STACK_SIZE - 16];
         }
@@ -340,7 +340,7 @@ void schedule(void) {
         if (tasks[next].cr3 != 0 && next != 0) {
             
             switch_cr3(tasks[next].cr3);
-            uint32_t kstack_top = tasks[next].kernel_stack_top;
+            uintptr_t kstack_top = tasks[next].kernel_stack_top;
             if (kstack_top == 0) {
                 kstack_top = (addr_t)&kernel_stacks[next][KERNEL_STACK_SIZE - 16];
             }
